+++ source/brailleDisplayDrivers/brailliantB.py
@@ -402,7 +402,7 @@ class InputGesture(braille.BrailleDisplayGesture, brailleInput.BrailleInputGestu
 					self.space = False
 			if key >= FIRST_ROUTING_KEY:
 				names.append("routing")
-				self.routingIndex = key - FIRST_ROUTING_KEY
+				self.cellIndexes = [key - FIRST_ROUTING_KEY]
 			else:
 				try:
 					names.append(KEY_NAMES[key])